+++ src/tools/users.js
@@ -6,8 +6,8 @@ import { z } from 'zod';
         name: "list_users",
         description: "List users in Zendesk",
         schema: {
-          page: z.number().optional().describe("Page number for pagination"),
-          per_page: z.number().optional().describe("Number of users per page (max 100)"),
+          page: z.number().int().optional().describe("Page number for pagination"),
+          per_page: z.number().int().optional().describe("Number of users per page (max 100)"),
           role: z.enum(["end-user", "agent", "admin"]).optional().describe("Filter users by role")
         },
         handler: async ({ page, per_page, role }) => {
@@ -32,7 +32,7 @@ import { z } from 'zod';
         name: "get_user",
         description: "Get a specific user by ID",
         schema: {
-          id: z.number().describe("User ID")
+          id: z.number().int().describe("User ID")
         },
         handler: async ({ id }) => {
           try {
@@ -59,7 +59,7 @@ import { z } from 'zod';
           email: z.string().email().describe("User's email address"),
           role: z.enum(["end-user", "agent", "admin"]).optional().describe("User's role"),
           phone: z.string().optional().describe("User's phone number"),
-          organization_id: z.number().optional().describe("ID of the user's organization"),
+          organization_id: z.number().int().optional().describe("ID of the user's organization"),
           tags: z.array(z.string()).optional().describe("Tags for the user"),
           notes: z.string().optional().describe("Notes about the user")
         },
@@ -94,12 +94,12 @@ import { z } from 'zod';
         name: "update_user",
         description: "Update an existing user",
         schema: {
-          id: z.number().describe("User ID to update"),
+          id: z.number().int().describe("User ID to update"),
           name: z.string().optional().describe("Updated user's name"),
           email: z.string().email().optional().describe("Updated email address"),
           role: z.enum(["end-user", "agent", "admin"]).optional().describe("Updated user's role"),
           phone: z.string().optional().describe("Updated phone number"),
-          organization_id: z.number().optional().describe("Updated organization ID"),
+          organization_id: z.number().int().optional().describe("Updated organization ID"),
           tags: z.array(z.string()).optional().describe("Updated tags for the user"),
           notes: z.string().optional().describe("Updated notes about the user")
         },
@@ -137,7 +137,7 @@ import { z } from 'zod';
         name: "delete_user",
         description: "Delete a user",
         schema: {
-          id: z.number().describe("User ID to delete")
+          id: z.number().int().describe("User ID to delete")
         },
         handler: async ({ id }) => {
           try {
